all: Implement FileContractRevisions
Originally I had the bright idea that we didn't need a separate
"revision" transaction -- we could submit the final revision and the
storage proof together in one transaction. After thinking about it
further, though, this makes no sense. Firstly, the renter needs a
window in which it can "contest" the revision submitted by the host.
Second, the final revision needs to be submitted before the proof
window opens; otherwise, the host could choose a revision that is
more favorable for the given segmentIndex.