all: Implement spend policies
Spend policies are a more generic form of Sia's "UnlockConditions",
based on Blockstream's "miniscript". They enable arbitrarily-complex
boolean expressions composed of public key requirements and timelocks,
but have no loops and are not Turing-complete.

The biggest missing feature here is DoS protection -- it's probably
really easy to construct a policy that decodes into something enormous,
and nothing guards against this. I'll probably modify DecodePolicy to
have some global limit on the number of sub-policies or something.