all: Refactor for StateElements
Basically:

  OutputID      -> ElementID
  SiacoinOutput -> SiacoinElement
  SiafundOutput -> SiafundElement
  FileContract  -> FileContractElement
  Beneficiary   -> SiacoinOutput, SiafundOutput

The FooElement types embed a generic StateElement type, along with
the transaction object they were derived from. The simplest example
is FileContractElement, which is literally just:

  type FileContractElement {
      StateElement
      FileContract
  }

StateElement contains an ElementID, a leaf index, and a Merkle proof.
It's fairly similar to the stateObject type in consensus (which now
also embeds StateElement).

Overall, this refactor is important because it unifies the various
"output" types by what they really have in common: they are all
present in the state accumulator. "Element" seemed like the best
descriptor for that; I also considered "Object", but...ugh. The
refactor also frees up the "Output" type names, so they now match
their old Sia v1 equivalents, which is nice.