consensus: Replace 64-bit accumulator flags with spent bool
Ultimately, I was not able to justify using flags instead of a bool.
There is little reason to distinguish between valid and missed file
contracts this way, since anyone who cares about whether a contract
resolved valid or not can simply look for a valid resolution in the
chain history. We were also recomputing proofs multiple times with
different flags in order to produce "better" error messages, but in
retrospect this was mostly wasted effort.

If we later add some other type of state element, and it turns out
that having generic flags would really benefit them, I'll be pretty
miffed.