mux: Prevent race condition via aliased write buffer
(Stream).Write was passing its buffer directly to (Mux).bufferFrame,
which stored it in m.write.payload before waking up (Mux).writeLoop.
This meant that Write could return before writeLoop actually flushed
the buffer, leading to a race condition.

Fortunately, the fix is easy: copy the buffer. To minimize memory
pressure, we reuse the existing capacity of m.write.payload instead
of allocating a new copy for each bufferFrame call. Usually, this
would be questionable, since it means we'd be keeping a potentially-
enormous buffer alive forever; but in this case, the buffer is
bounded by the maximum frame size, so it will never grow larger than
a kilobyte or so.