consensus,types: Tweak SigHash and (Transaction).ID
ID was hashing the transaction's Merkle proofs (and leaf indices),
which meant that the IDs of transactions in the txpool were changing
after every applied block. Now that information is omitted, which
introduces some malleability -- but any modifications to a transaction
that don't change its ID, *should* cause it to become invalid (aside
from the standard Merkle proof updates, of course). In the same vein,
I added length prefixes to the hash to prevent trivial ID collisions
(e.g. placing an encoded []SiacoinOutput in the ArbitraryData).