consensus: Use Work-based PoW in v2
Pre-hardfork, the Work values are computed as the inverse of their
Target counterparts; post-hardfork, it's the opposite. Importantly,
this means that PoW is still checked by comparing against ChildTarget,
even though all of the difficulty adjustment math is in terms of Work.
Other Target-based fields, like Depth and OakTarget, should be ignored
post-hardfork.

It would be nice if we could check work by comparing the difficulty to
the expected number of hashes for the ID, but this would break compat-
ibility with existing miners (and introduce a costly 256-bit division
in the mining loop -- no thanks).

So why make this change at all? Because Work-based math is more intuitive
and easier for newcomers to understand; also, future nodes, ignorant of
pre-v2 history, will be able to strip out the Target math entirely.