types: Change v2 address algorithm
This makes threshold policy addresses more like a Merkle root;
the address is a hash of the addresses of the sub-policies.
Furthermore, you can replace any sub-policy with an "opaque"
version that has the same hash, but reveals no information.
This is good for privacy; it also saves some space in policies
with big, unused sub-policies. But that's actually not the
primary motivation here. The primary motivation is simplifying
validation!

When we validate a pubkey spend policy, we check it against
*every* signature provided alongside the input. If we didn't --
if we instead only checked the *next* signature -- we would
reject some inputs with valid signatures. Consider:

  thresh(1, [thresh(1, [pk(A), pk(B)]), pk(C)])

i.e. "either A or B or C." Now imagine we have an input with
one signature from key C. Clearly this should be valid. But
if we only checked the *next* signature, we would reject it:
first we would check against A; after that failed, we would
discard the signature, and never reach C.

Another strategy would be to only discard after hitting a
valid signature. That way, the signature from C would be
checked against A, B, and C, ultimately succeeding. But this
approach fails on a different case:

  thresh(1, [above(3), pk(A), pk(B)])

Now, which signatures should we provide? It depends on
whether the height is above 3. If so, we should provide a
signature from A; otherwise, a signature from B.

This isn't the end of the world -- changes in height are
allowed to invalidate transactions, after all -- but it
does feel a bit user-hostile. Plus, it's wasteful: the
signature is checked three times instead of once. The new
approach fixes these problem by having the transaction
signer specify which sub-policies they intend to satisfy.
They do this by replacing the unsatisfied sub-policies
with "opaque" policies. The validator can then use a much
simpler strategy for handling signatures: the slice of
signatures provided in the input should exactly match a
depth-first traversal of the policy. Each signature only
needs to be checked once, and failure to verify can be
treated as a fatal error.

Anyway, since this stuff is only relevant to complex
policies, I was careful to ensure that the standard single-
pubkey policy is still simple and cheap to encode -- it
should only require one hash. (In fact, the algorithm
for non-threshold policies hasn't changed at all.)

One last note: since the opaque hash is actually the same
as the policy's address, there is potential privacy leak:
if you use the same keypair in two different policies, then
anyone can tell that they are connected, even if one of
them is "opaque." This doesn't seem like a big deal; you
shouldn't be reusing keypairs like that anyway, and if you
*really* want to, you can always wrap it in a thresh(1) or
an above(0) or something to change the opaque hash.