types,consensus: Expose txn output IDs
For some reason I thought these could be kept internal to consensus,
but in practice you need them whenever you want to spend an ephemeral
output.

One quirk is that some of the methods are "pointlessly" defined on
V2Transaction, despite not using the transaction data -- they only use
the ID. Why not make these methods of TransactionID? Because then both
Transaction and TransactionID would have a SiacoinOutputID method, and
you'd Just Have To Know which one to use in which circumstance. Better
to forcibly associate v1 IDs with v1 txns and v2 IDs with v2 txns.