consensus,chain: Add ForEachTreeNode
Previously, the DB only stored tree nodes that were present in the
proofs of added or updated elements. This is not sufficient: there
may be nodes that are created by the addition of new elements, but
do not appear in any proofs. Fortunately, we do compute and store
these nodes in treeGrowth.

I went with a "row,col" API for uniquely identifying nodes in the
tree. Surprisingly, I've never done this before in any Merkle tree-
related code. It feels a bit clunky to me, particularly the frequent
need to xor the col to get its sister, but it works well enough.