consensus: Allow v1 contracts to end after v2 hardfork
Unfortunately this can't be a consensus rule because there are
contracts already in existence that violate it, e.g.

  fbc9bd9adb100755a6cedc8d9e38d46aa2931cf719dc866e0f50b8855b0cffe0

which expires at height 4,087,948. Such contracts will simply
become unresolvable after the v2 hardfork. We will add checks to
renterd and hostd that guard against users accidentally forming
v1 contracts that would be unresolvable.