consensus: Move commitment check to ValidateBlock
We can't validate the commitment in ValidateOrphan because our
State might be slightly inaccurate: specifically, its Elements
accumulator might not have been updated. This happens when we're
validating a sidechain pre-v2, since we're storing the full tree
and only updating it *after* validating each of the sidechain
blocks as orphans.