types: Implement new multiproof encoding
With some clever bit twiddling (one of life's greatest joys), we
can recover all of the Merkle proof sizes from the element's leaf
index and the total number of leaves, rather than encoding each
size as a uint8 like before. Even better, the caller doesn't need
to pass numLeaves to the encoder; it can be inferred (at least, to
a sufficiently-close approximation) from the supplied elements!