consensus: Require valid signatures on V2FileContractRenewal.NewContract
Unlike the FinalRevision, there is no danger of allowing the new contract
to be independently valid, as it cannot be broadcast without also being
funded. Note that this is already possible with any other v2 contract
formation: anyone can "replay" the contract formation in a transaction of
their own, but they need to supply the contract funding themselves.